# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ekai Support Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the current portal user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalUser"
        "403":
          description: Authenticated with Clerk but not invited to the portal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /invites/accept:
    post:
      operationId: acceptInvite
      tags: [auth]
      summary: Accept an invite token and join the portal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteAcceptance"
      responses:
        "200":
          description: Invite accepted; portal user created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalUser"
        "400":
          description: Invalid or expired invite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /invites/preview:
    get:
      operationId: previewInvite
      tags: [auth]
      summary: Preview invite details by token (pre-auth)
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Invite details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitePreview"
        "400":
          description: Invalid or expired invite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      properties:
        message: { type: string }
        code: { type: ["string", "null"] }
      required: [message]
    PortalUser:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        role: { type: string, enum: [customer, ekai_agent, admin] }
        orgId: { type: ["integer", "null"] }
        orgName: { type: ["string", "null"] }
        active: { type: boolean }
        createdAt: { type: string }
        lastLogin: { type: ["string", "null"] }
      required: [id, email, name, role, orgId, orgName, active, createdAt, lastLogin]
...
    InviteAcceptance:
      type: object
      properties:
        token: { type: string, minLength: 1 }
      required: [token]
    InvitePreview:
      type: object
      properties:
        email: { type: string }
        role: { type: string, enum: [customer, ekai_agent, admin] }
        orgName: { type: ["string", "null"] }
        expiresAt: { type: string }
      required: [email, role, orgName, expiresAt]
...